Split HTML renderer into focused submodules
Replace the 481-line src/graph_view/render_html.rs with a directory
containing one module per concern: escape, layout, style, svg, outline,
noscript. The directory's mod.rs (113 lines) holds only the HtmlRenderer
adapter and orchestrates the four output sections. Every submodule sits
well under the AGENTS.md 400-line cap, making future renderer work
easier to review and isolate.

Behaviour is unchanged; the existing render_html_tests.rs becomes
render_html/tests.rs.
